requirements.txt
# Основные зависимости FastAPI
fastapi==0.115.0
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
pydantic==2.10.3
python-multipart==0.0.20

# Google Calendar API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Microsoft Graph API
msal==1.25.0
requests==2.31.0

# Конфигурация и утилиты
python-dotenv==1.1.1

# Продакшен зависимости
gunicorn==21.2.0

# Дополнительные утилиты
click==8.1.7
colorama==0.4.6

# Для разработки (опционально)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

